requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1 
streamlit==1.30.0
boto3==1.34.14
python-dotenv==1.0.0 